backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pandas
scikit-learn
